+++ daisy_ui_components_site/assets/js/storybook.js
@@ -1,4 +1,4 @@
-document.documentElement.dataset.theme = 'light';
+document.documentElement.dataset.theme = "light";
 
 // If your components require any hooks or custom uploaders, or if your pages
 // require connect parameters, uncomment the following lines and declare them as
@@ -12,10 +12,9 @@ document.documentElement.dataset.theme = 'light';
 //   window.storybook = { Hooks, Params, Uploaders };
 // })();
 
-
 // If your components require alpinejs, you'll need to start
 // alpine after the DOM is loaded and pass in an onBeforeElUpdated
-// 
+//
 // import Alpine from 'alpinejs'
 // window.Alpine = Alpine
 // document.addEventListener('DOMContentLoaded', () => {