+++ daisy_ui_components_site/storybook/_root.index.exs
@@ -5,12 +5,19 @@ defmodule Storybook.Root do
   use PhoenixStorybook.Index
 
   def folder_icon, do: {:local, "hero-book-open", "psb-w-4 psb-h-4"}
-  def folder_name, do: "DaisyUI"
+  def folder_name, do: "🌻 DaisyUIComponents"
 
   def entry("welcome") do
     [
       name: "Welcome",
       icon: {:local, "hero-home", "psb-w-4 psb-h-4"}
+    ]
+  end
+
+  def entry("setup") do
+    [
+      name: "Setup",
+      icon: {:local, "hero-adjustments-horizontal", "psb-w-4 psb-h-4"}
     ]
   end
 end