+++ lib/daisy_ui_components/swap.ex
@@ -27,6 +27,8 @@ defmodule DaisyUIComponents.Swap do
     attr :name, :string, required: true
   end
 
+  slot :inner_block
+
   def swap(assigns) do
     assigns =
       assigns
@@ -34,7 +36,11 @@ defmodule DaisyUIComponents.Swap do
 
     ~H"""
     <label class={@class} {@rest}>
-      <input class="hidden" type="checkbox" />
+      <%= if render?(@inner_block) do %>
+        {render_slot(@inner_block)}
+      <% else %>
+        <input class="hidden" type="checkbox" />
+      <% end %>
       <.swap_mode
         :for={swap_on <- @swap_on}
         type={Map.get(swap_on, :type, "label")}